pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "companion-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "pgvector>=0.3.0",
]

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["**/*.txt", "**/*.md", "**/*.jinja", "**/*.json"]

[tool.uvicorn]
factory = false
